separate stdio and http server code
